requirements.txt
absl-py==0.9.0
aioconsole==0.1.16
aiohttp==3.6.2
alabaster==0.7.12
altgraph==0.17
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
anime==1.1.0
apipkg==1.5
appdirs==1.4.3
asgiref==3.2.5
asn1crypto==1.0.1
astor==0.8.1
astroid==2.3.1
astropy==3.2.1
async-timeout==3.0.1
asyncio==3.4.3
atomicwrites==1.3.0
attrs==19.2.0
auto-py-to-exe==2.7.5
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.0
better-exceptions==0.2.2
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.0
bokeh==1.3.4
boto==2.49.0
bottle==0.12.18
bottle-websocket==0.2.9
Bottleneck==1.2.1
brewer2mpl==1.4.1
cachetools==4.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
clipboard==0.0.4
cloudpickle==1.2.2
cloudscraper==1.2.36
clyent==1.2.2
colorama==0.4.1
coloredlogs==14.0
colour==0.1.5
comtypes==1.1.7
conda==4.8.2
conda-build==3.18.9
conda-package-handling==1.6.0
conda-verify==3.4.2
configparser==5.0.0
contextlib2==0.6.0
COVID19Py==0.3.0
crayons==0.3.0
cryptography==2.7
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.2
decorating==0.6.1
decorator==4.4.0
defusedxml==0.6.0
discord==1.0.1
discord.py==1.3.3
distlib==0.3.0
distributed==2.5.2
dj-database-url==0.5.0
Django==3.0.4
django-bootstrap4==2.0.1
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-markup==1.4.1
django-pagedown==2.1.2
docutils==0.15.2
Eel==0.12.4
entrypoints==0.3
et-xmlfile==1.0.1
execnet==1.7.1
fastcache==1.1.0
ffmpeg-python==0.2.0
filelock==3.0.12
Flask==1.1.1
fsspec==0.5.2
future==0.17.1
gast==0.2.2
gevent==1.4.0
gevent-websocket==0.10.1
glob2==0.7
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.9.0
HeapDict==1.0.1
html5lib==1.0.1
humanfriendly==8.2
idna==2.6
imageio==2.6.0
imagesize==1.1.0
importlib-metadata==1.6.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyter-manim==1.1
jupyterlab==1.1.4
jupyterlab-server==1.0.6
jupyterthemes==0.20.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.0
Kivy==1.11.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
lesscpy==0.14.0
libarchive-c==2.8
llvmlite==0.29.0
locket==0.2.0
logzero==1.5.0
lxml==4.3.4
lyricwikia==0.1.11
mal==0.4.0
manga-py==1.19.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multidict==4.7.5
multipledispatch==0.6.0
music-dl==0.2.1
music21==5.7.2
mutagen==1.44.0
mysql==0.0.2
mysqlclient==1.4.6
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.15.0
numpydoc==0.9.1
oauthlib==3.1.0
olefile==0.46
opencv-python==3.4.2.17
openpyxl==3.0.0
opt-einsum==3.2.1
packaging==20.3
pafy==0.5.5
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib==1.0.1
pathlib2==2.3.5
patsy==0.5.1
pefile==2019.4.18
pep8==1.7.1
pickleshare==0.7.5
Pillow==5.2.0
pkginfo==1.5.0.1
pluggy==0.13.0
ply==3.11
prettyplotlib==0.1.7
progressbar==2.5
progressbar2==3.50.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.3
psutil==5.6.3
psycopg2==2.8.5
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycairo==1.19.1
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.9.7
pycurl==7.43.0.3
pydub==0.23.0
PyExecJS==1.5.1
pyflakes==2.1.1
pygame==1.9.6
Pygments==2.4.2
PyInstaller==3.6
pylint==2.4.2
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyperclip==1.8.0
pyreadline==2.1
pyrsistent==0.15.4
PySocks==1.7.1
pytesseract==0.3.3
pytest==5.2.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-forked==1.1.3
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
pytest-xdist==1.32.0
python-dateutil==2.8.0
python-magic==0.4.15
python-mal==0.1.7
python-utils==2.4.0
python3-mal==0.2.20
pytube3==9.6.4
pytz==2019.3
PyWavelets==1.0.3
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
requests==2.18.4
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rope==0.14.0
rsa==4.0
ruamel-yaml==0.15.46
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.1.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spotdl==2.1.0
spotipy==2.12.0
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.9
sqlparse==0.3.1
statsmodels==0.10.1
sympy==1.4
tables==3.5.2
taichi==0.6.10
tblib==1.4.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
titlecase==0.12.0
tldextract==2.2.2
toolz==0.10.0
tornado==6.0.3
tqdm==4.47.0
traitlets==4.3.3
typing-extensions==3.7.4.2
unicode-slugify==0.1.3
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.22
virtualenv==20.0.10
wcwidth==0.1.7
webdriver-manager==2.4.0
webencodings==0.5.1
websockets==8.1
Werkzeug==0.16.0
whichcraft==0.6.1
whitenoise==5.1.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.1
xlwings==0.15.10
xlwt==1.3.0
yarl==1.4.2
youtube-dl==2020.3.24
zict==1.0.0
zipp==0.6.0